+++ CHANGELOG.md
@@ -7,6 +7,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 
 * Bugfix - User Invite - New User Role preselected fix
 * Bugfix - App User Management - Confirmation and User Role Changes
+* Feature - User Management - fetch roles (api refactoring)
 
 
 ## 0.7.0